Get auto-buy working
buy_value wasn't calculated correctly (wasn't dividing by 100) and
I got a quote precision error so I needed to round_de(sym.quote_precision).
But the quote_precision needs to be a u32 not a u64. So I had to add
the de_string_or_number_to_i32 and u32 and change the various _precision
value in struct Symbol. Fixed the output.

Also, added checking for asset != quote_asset so we ignore "USDUSD" and
the like. Not sure why I didn't run into that before. Added rounding
sell_qty to quote_precision in auto-sell as I did with buy_value. And
some cleanup of the output.
quote_precision in the

+++ src/binance_exchange_info.rs
@@ -7,7 +7,9 @@ use rust_decimal::prelude::*;
 use std::collections::HashMap;
 
 use crate::common::get_req_get_response;
+//use crate::de_string_or_number::u32_de_string_or_number;
 use crate::de_string_or_number::de_string_or_number_to_u64;
+use crate::de_string_or_number::de_string_or_number_to_u32;
 use crate::{common::OrderType, configuration::Configuration};
 
 use strum_macros::IntoStaticStr;
@@ -242,22 +244,22 @@ impl SymbolFilters {
 pub struct Symbol {
     pub symbol: String,     // +enum BTCUSD?
     pub base_asset: String, // +enum BTC?
-    #[serde(deserialize_with = "de_string_or_number_to_u64")]
-    pub base_asset_precision: u64,
-    #[serde(deserialize_with = "de_string_or_number_to_u64")]
-    pub base_commission_precision: u64,
+    #[serde(deserialize_with = "de_string_or_number_to_u32")]
+    pub base_asset_precision: u32,
+    #[serde(deserialize_with = "de_string_or_number_to_u32")]
+    pub base_commission_precision: u32,
     pub iceberg_allowed: bool,
     pub is_margin_trading_allowed: bool,
     pub is_spot_trading_allowed: bool,
     pub oco_allowed: bool,
     pub quote_asset: String, // +enum USD?
-    #[serde(deserialize_with = "de_string_or_number_to_u64")]
-    pub quote_asset_precision: u64,
-    #[serde(deserialize_with = "de_string_or_number_to_u64")]
-    pub quote_commission_precision: u64,
+    #[serde(deserialize_with = "de_string_or_number_to_u32")]
+    pub quote_asset_precision: u32,
+    #[serde(deserialize_with = "de_string_or_number_to_u32")]
+    pub quote_commission_precision: u32,
     pub quote_order_qty_market_allowed: bool,
-    #[serde(deserialize_with = "de_string_or_number_to_u64")]
-    pub quote_precision: u64,
+    #[serde(deserialize_with = "de_string_or_number_to_u32")]
+    pub quote_precision: u32,
     pub status: String, // +enum TRADING?
     pub permissions: Vec<String>,
     pub order_types: Vec<OrderType>, // +HashSet<OrderType>?
